+++ .eslintrc.js
@@ -11,6 +11,9 @@ module.exports = {
   plugins: [
     '@typescript-eslint',
   ],
+  extends: [
+    'plugin:@next/next/recommended',
+  ],
   ignorePatterns: '*.js',
   overrides: [
     {