+++ packages/abjad-web/src/app/page.tsx
@@ -1,17 +1,65 @@
 'use client'
 
 import { useState } from 'react'
+import Select from 'react-select'
 import { Abjad, convert } from 'abjad-convert'
 
 export default function Home() {
 	const [textBoxValue, setTextBoxValue] = useState('')
 	const [resultText, setResultText] = useState('')
+	const [fromValue, setFromValue] = useState<Abjad | undefined>(Abjad.Arabic)
+	const [toValue, setToValue] = useState<Abjad | undefined>(Abjad.ImperialAramaic)
 
+	const options = [
+		{ value: Abjad.Arabic, label: 'Arabic ابجد' },
+		{ value: Abjad.ImperialAramaic, label: 'Imperial Aramaic 𐡀𐡁𐡂𐡃' },
+		{ value: Abjad.Phoenician, label: 'Phoenician 𐤀𐤁𐤂𐤃' },
+		{ value: Abjad.Syriac, label: 'Syriac ܐܒܓܕ' },
+		{ value: Abjad.Tifinagh, label: 'Tifinagh ⴰⴱⵊⴷ' },
+		{ value: Abjad.Ugaritic, label: 'Ugaritic 𐎀𐎁𐎂𐎄' },
+	]
+	const FromSelect = () => <div style={{ marginBottom: '10px' }}>
+		<label htmlFor="fromDropdown" style={{ marginRight: '10px' }}>
+			From:
+		</label>
+		<Select
+			id="fromDropdown"
+			isSearchable={false}
+			options={options}
+			defaultValue={options[0]}
+			onChange={(selectedOption) => {
+				setFromValue(selectedOption?.value)
+				if (!selectedOption?.value || !toValue) {
+					return
+				}
+				const result = convert(textBoxValue, selectedOption?.value, toValue)
+				setResultText(result)
+			}}
+		/>
+	</div>
+	const ToSelect = () => <div style={{ marginBottom: '10px' }}>
+		<label htmlFor="toDropdown" style={{ marginRight: '10px' }}>
+			To:
+		</label>
+		<Select
+			id="toDropdown"
+			isSearchable={false}
+			options={options}
+			defaultValue={options[1]}
+			onChange={(selectedOption) => {
+				setToValue(selectedOption?.value)
+				if (!fromValue || !selectedOption?.value) {
+					return
+				}
+				const result = convert(textBoxValue, fromValue, selectedOption?.value)
+				setResultText(result)
+			}}
+		/>
+	</div>
 	const handleConvert = () => {
-		const fromDropdown = document.getElementById('fromDropdown') as HTMLSelectElement
-		const fromValue = fromDropdown.value as Abjad
-		const toDropdown = document.getElementById('toDropdown') as HTMLSelectElement
-		const toValue = toDropdown.value as Abjad
+		if (!fromValue || !toValue) {
+			return
+		}
 		const result = convert(textBoxValue, fromValue, toValue)
 		setResultText(result)
 	}
@@ -19,65 +67,36 @@ export default function Home() {
 	return (
 		<main style={{ textAlign: 'center', padding: '20px' }}>
 			<h1>Abjad Converter</h1>
-
-			<div style={{ marginBottom: '10px' }}>
-				<label htmlFor="fromDropdown" style={{ marginRight: '10px' }}>
-					From:
-				</label>
-				<select id="fromDropdown">
-					<option value={Abjad.Arabic}>Arabic</option>
-					<option value={Abjad.ImperialAramaic}>Imperial Aramaic</option>
-					<option value={Abjad.Phoenician}>Phoenician</option>
-					<option value={Abjad.Syriac}>Syriac</option>
-					<option value={Abjad.Tifinagh}>Tifinagh</option>
-					<option value={Abjad.Ugaritic}>Ugaritic</option>
-				</select>
-			</div>
-
-			<div style={{ marginBottom: '10px' }}>
-				<label htmlFor="toDropdown" style={{ marginRight: '10px' }}>
-					To:
-				</label>
-				<select id="toDropdown">
-					<option value={Abjad.Arabic}>Arabic</option>
-					<option value={Abjad.ImperialAramaic}>Imperial Aramaic</option>
-					<option value={Abjad.Phoenician}>Phoenician</option>
-					<option value={Abjad.Syriac}>Syriac</option>
-					<option value={Abjad.Tifinagh}>Tifinagh</option>
-					<option value={Abjad.Ugaritic}>Ugaritic</option>
-				</select>
-			</div>
-
-			<div style={{ marginBottom: '10px' }}>
-				<label htmlFor="editTextBox" style={{ marginRight: '10px' }}>
-					Enter Text:
-				</label>
-				<textarea
-					id="editTextBox"
-					placeholder="Type here"
-					value={textBoxValue}
-					onChange={(e) => setTextBoxValue(e.target.value)}
-					style={{ padding: '10px', width: '100%', minHeight: '100px' }}
-				/>
-			</div>
-
-			<div style={{ marginBottom: '10px' }}>
-				<button onClick={handleConvert} style={{
-					padding: '10px',
-					backgroundColor: '#4CAF50',
-					color: 'white',
-					border: 'none',
-					borderRadius: '5px'
-				}}>
-					Convert
-				</button>
-			</div>
-
+			{FromSelect()}
+			{ToSelect()}
+			<label htmlFor="editTextBox" style={{ marginRight: '10px' }}>
+				Enter Text:
+			</label>
+			<textarea
+				id="editTextBox"
+				placeholder="Type here"
+				value={textBoxValue}
+				onChange={(e) => {
+					setTextBoxValue(e.target.value)
+					if (!fromValue || !toValue) {
+						return
+					}
+					const result = convert(e.target.value, fromValue, toValue)
+					setResultText(result)
+				}}
+				style={{ padding: '10px', width: '100%', minHeight: '100px', fontSize: '25px' }}
+			/>
 			<div>
 				<label htmlFor="resultLabel" style={{ marginRight: '10px' }}>
 					Result:
 				</label>
-				<span id="resultLabel" style={{ fontWeight: 'bold' }}>{resultText}</span>
+				<br/>
+				<span
+					id="resultLabel"
+					style={{ padding: '10px', width: '100%', minHeight: '100px', fontSize: '25px' }}
+				>
+					{resultText}
+				</span>
 			</div>
 		</main>
 	)